requirements.txt
 Django
psycopg2-binary
pandas
openpyxl
